Move unit tests to a subfolder
