+++ test/unit/algebra/Algebra.tests.js
@@ -1,4 +1,4 @@
-import { linearCombination } from '../../src/algebra/Algebra.js';
+import { linearCombination } from '../../../src/algebra/Algebra.js';
 
 describe( "Algebra", function () {
